+++ api/azure_storage_api.py
@@ -4,18 +4,17 @@ import hashlib
 import datetime
 import numpy as np
 from azure.storage.blob import BlobServiceClient
-# from custom_exceptions import (
-#     ConnectionStringError,
-#     MountContainerError,
-#     GetBlobError,
-#     UploadImageError,
-#     UploadInferenceResultError,
-#     GetFolderUUIDError,
-#     FolderListError,
-#     GenerateHashError,
-#     ProcessInferenceResultError,
-# )
-from . import custom_exceptions as ce
+from custom_exceptions import (
+    ConnectionStringError,
+    MountContainerError,
+    GetBlobError,
+    UploadImageError,
+    UploadInferenceResultError,
+    GetFolderUUIDError,
+    FolderListError,
+    GenerateHashError,
+    ProcessInferenceResultError,
+)
 """
 ---- user-container based structure -----
 - container name is user id
@@ -36,7 +35,7 @@ async def generate_hash(image):
         hash = hashlib.sha256(image).hexdigest()
         return hash
 
-    except ce.GenerateHashError as error:
+    except GenerateHashError as error:
         print(error)
 
 
@@ -60,9 +59,9 @@ async def mount_container(connection_string, container_uuid, create_container=Tr
                 container_client = blob_service_client.create_container(container_name)
                 return container_client
         else:
-            raise ce.ConnectionStringError("Invalid connection string")
+            raise ConnectionStringError("Invalid connection string")
 
-    except ce.MountContainerError as error:
+    except MountContainerError as error:
         print(error)
         return False
 
@@ -77,7 +76,7 @@ async def get_blob(container_client, blob_name):
         blob_content = blob.readall()
         return blob_content
 
-    except ce.GetBlobError as error:
+    except GetBlobError as error:
         print(error)
         return False
 
@@ -110,7 +109,7 @@ async def upload_image(container_client, folder_name, image, hash_value):
             container_client.upload_blob(blob_name, image, overwrite=True)
             return blob_name
 
-    except ce.UploadImageError as error:
+    except UploadImageError as error:
         print(error)
         return False
 
@@ -127,7 +126,7 @@ async def upload_inference_result(container_client, folder_name, result, hash_va
             container_client.upload_blob(json_name, result, overwrite=True)
             return True
 
-    except ce.UploadInferenceResultError as error:
+    except UploadInferenceResultError as error:
         print(error)
         return False
 
@@ -152,7 +151,7 @@ async def get_folder_uuid(container_client, folder_name):
                     if folder_json["folder_name"] == folder_name:
                         return blob.name.split(".")[0].split("/")[-1]
         return False
-    except ce.GetFolderUUIDError as error:
+    except GetFolderUUIDError as error:
         print(error)
         return False
 
@@ -176,7 +175,7 @@ async def folder_list(container_client):
                     folder_list.append(folder_json["folder_name"])
         folder_list.sort()
         return list(set(folder_list))
-    except ce.FolderListError as error:
+    except FolderListError as error:
         print(error)
         return []
 
@@ -240,6 +239,6 @@ async def process_inference_results(data, imageDims):
         # add totalBoxes attribute to the inference results
         data[0]["totalBoxes"] = sum(1 for box in data[0]["boxes"])
         return data
-    except ce.ProcessInferenceResultError as error:
+    except ProcessInferenceResultError as error:
         print(error)
         return False
