+++ custom_exceptions.py
